LayoutTests/fast/js/dfg-arguments-osr-exit-expected.txt

Tests aliased uses of 'arguments' that require reification of the Arguments object on OSR exit.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is 74
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS bar(42) is "4232"
PASS successfullyParsed is true

TEST COMPLETE
